Non-Markovian model added. The probability of eos is constant to make the computation of the exact distributions easier.
Tests of the variance and a Kolmogorov-Smirnov test for the distribution has been added. Asserts have been computed for n=250. If we want the test to run faster, we can reduce this number.

All tests have been run with n=1000000 to verify the monte carlo error goes to zero.

All tests have been run with multiple seeds.